In tree2Paths, check for tree concordance, change binarize behavior, and update documentation
This update is intended to address the following issues:

 - The previous version of tree2Paths exhibited issues when the tree topology of the phenotype
   tree differed from that of the masterTree in the supplied treesObject.

 - The behavior of 'binarize' in tree2Paths has been variously set to be optimal for binary
   or continuous phenotype trees, rather than flexibly choosing behavior based on the type of
   data within the phenotype tree.

 - The documentation of tree2Paths needed to be updated to reflect these changes and improve clarity.

tree2Paths will now fail if the tree topology of the phenotype tree is discordant from that
of the masterTree within the supplied treesObject, and will report an error from the test for
concordance using 'all.equal' in ape.

tree2Paths will now by default binarize paths for binary phenotype trees and will not binarize
paths for continuous phenotype trees.