Create binarize default and accurately compare topologies in tree2Paths
This update implements the following suggested fixes suggested by @raghavendranpartha
from the pull request for 'ModifyPathCreation':

- Uses is.null(binarize) to force setting a default for 'binarize'

- Checks topologies by correctly specifying the masterTree within treesObj

It does not implement the following suggestion:

- Changing from all.equal.phylo() to base::all.equal to check topologies

The 'ape' implementation of all.equal() specifically checks for concordant topologies,
so it should be preferable to the base check of object equality.

I have checked that this version succeeds for a binary phenotype tree created using
the master tree topology and fails when the binary phenotype tree has a different topology
(specifically when it is a pruned tree).