Update binary vignette and associated runMarineSub script
The version of the binary vignette currently on master contains references
to some old data files and no longer recommended practices for implementing
RERconverge. This is a (to-be-continued) update to both the .Rmd file
and the script on which the vignette is based to bring the walk-through up
to date.