Update vignette with options for foreground2Tree
Changed the binary vignette to provide detailed explanations and visualization
of the options for `foreground2Tree` (ancestral, terminal, all, and weighted).
This should show the user how the various options result in different branch
lengths on the binary tree, using the marine clades as an example.

The vignette still includes the option to use the interactive (Rshiny)
`selectforegroundbranches` function. If this won't be included with the final
package, these need to be commented out. They are now separated from the
other options for generating trees, so excluding them won't eliminate
any crucial information.