Edit binary sections of combined vignette walkthrough
Made some preliminary changes to make the binary sections of the vignette
walkthrough compatible with the continuous sections. Also modified some
paths to allow users to access files in inst/extdata (the ones they will
also want to look at in a text editor).

The Rmd still requires some manual editing so that the user can source the
plottingFuncs.R script from the UpdateVignette branch; this version of the
script contains a functional `plotRers`, as well as the function
`plotTreeHighlightBranches`, which is used in the vignette.