Revised and knit full binary + continuous walk-through
These changes should make the binary and continuous sections of the walk-through compatible with
each other and should permit knitting on (almost) any machine. The updated walk-through is highly
detailed and provides information about specifying trees and running correlation functions with
both binary and continuous traits.

Remaining issues/things to fix:
-The species in the binary tree should (probably?) be made the same as those in the continuous tree.
Currently using all species because there is not yet a way to specify `useSpecies` in foreground2Tree.
-Others should verify that the necessary functions and datasets load on installation.
The current code chunks have a few workarounds that allow local sourcing and loading if necessary.
-The knitted html doesn't allow scrolling all the way through the toc because it is too long.
This will probably require modifications to custom.css.
-The walk-through references the `selectforegroundbranches` function relying on Rshiny, though it is
not currently distributed with the package. We should decide whether to include it.
-There is a reference missing for details about correlation methods for binary and continuous traits.
-Figures are HUGE. We should figure out how to make them a bit smaller (or at least reduce margins).
-The "numbers" 2a - 2d don't format correctly in the binary tree section.