Working towards a knittable version of the vignette
More changes to ensure that the vignette steps run and the Rmd file knits
on all computers. Commented out some code referencing local repo directories
and modified a few functions to help with export and dependencies. Also
updated species names in two files to make them compatible with the trees.
Remaining issue:
-plotTreeHighlightBranches prevents knitting. The function is set to export
in plottingFuncs and not within the if(F) loop. Still troubleshooting.