Removed check for topology match btw trait and master trees
This commit removes the topology equality check from tree2Paths,
in order to allow the vignette to run for the binary trait tree,
which is missing a species present in the master tree. This may
still be an issue for other cases where there is topological
discordance, unless a check is present in matchNodesInject.
See issue #22.