Revisions to error handling and pruning when trait and master trees not equivalent
Revised tree2Paths, foreground2Tree, and getAllCor to provide warnings when trees
don't match and to prune the trait tree to include only species in the master.

***Remaining issue***
tree2Paths does not prune the master tree, so matchAllNodes throws a topology
discordance error if the trait tree is a subset of the master tree.