Revisions to prevent NAs from being back-propagated in diffBranches
Changed code to allow user to set NA branches to zero when estimating
differences with upstream branches (so that NAs aren't propagated
throughout the tree). Also beginning to develop a quicker changeInRers.

Still need: a way to quickly assign matrix values based on function with
one output representing the column ids and another representing the values
to assign.