Combined updates to readTrees for master tree specification and reorientation
Revised readTrees in an attempt to resolve merge conflicts with ChangeInRers
branch. This branch contained an option to re-orient the master tree and all
sub-trees relative to a known or inferred outgroup. It may have been preesent
in another branch as well.

Note that this is not the same as the changes in the UserSpecMasterTree branch,
which makes rotation of the trees optional. I (Wynn) tentatively assert that the
current version in AddEstimateTreeFunctions is less error-prone.

Remaining to-do:
-remove the requirement of 'limma'; this is for ChangeInRers only (test to
confirm) and will require a Bioconductor install on the part of the user.