Removed limma requirement; added new plotting function
Eliminated require(limma) from RERfuncs. This was a holdover
from the ChangeInRers branch (it's required for new functions there).
limma is a BioConductor package, so it may cause issues for users
to install.

Also put new gg-based plotting function in the R directory.

To-do:
-change name of function in new plotting script, so that it doesn't
replace the default plotting function (allows plotting in either base
R or ggplot)